move dot_writer to seperate files
